+++ src/app/gui/params/params.component.ts
@@ -1,7 +1,8 @@
 import {Component, EventEmitter, OnInit, Output} from '@angular/core'
-import {DEFAULT_PARAMS} from "../../model/params"
 import {MatDialog} from '@angular/material/dialog'
 import {ShowCodeDialog} from "../show-code-dialog/show-code-dialog.component"
+import {Params} from "@angular/router"
+import {DEFAULT_PARAMS} from "../../model/params"
 
 @Component({
   selector: 'app-params',
@@ -10,25 +11,65 @@ import {ShowCodeDialog} from "../show-code-dialog/show-code-dialog.component"
 })
 export class ParamsComponent implements OnInit {
   @Output() paramEmitter = new EventEmitter<object>()
-  public params = DEFAULT_PARAMS
+  public params: Params = DEFAULT_PARAMS
+  public customColor = "red"
+  public customBackgroundColor = "red"
+  public customBackgroundImage: string
+  public fontsize: 20
 
   constructor(public dialog: MatDialog) {
   }
 
   ngOnInit(): void {
-    setTimeout(() => this.onUpdate())
+    this.restartAnimation()
   }
 
   onUpdate(): void {
-    this.paramEmitter.emit(this.params)
+    this.paramEmitter.emit(this.applyCustomParams(this.params))
+  }
+
+  applyCustomParams(params: Params): Params {
+    let updatedParams = {...params}
+    updatedParams.color = (this.params.color === 'customColor') ? this.customColor : this.params.color
+    switch (this.params.background) {
+      case 'customBackgroundColor':
+        updatedParams.background = this.customBackgroundColor
+        break
+      case 'customBackgroundImage':
+        updatedParams.background = `url('${this.customBackgroundImage}')`
+        break
+      default:
+        updatedParams.background = this.params.background
+    }
+    return updatedParams
+  }
+
+  onFileUpload(event: any) {
+    const file = event.target.files[0]
+    const reader = new FileReader()
+    reader.onload = () => {
+      // TODO: set canvas width and height to image width and height?
+      this.customBackgroundImage = reader.result as string
+      this.onUpdate()
+    }
+    reader.readAsDataURL(file)
+  }
+
+  restartAnimation(): void {
+    setTimeout(() => this.onUpdate())
   }
 
   openShowCodeDialog(): void {
+    const paramsToDisplay = this.applyCustomParams(this.params)
+    if (paramsToDisplay.background.includes('url')) {
+      paramsToDisplay.background = 'url(path/to/your/image)'
+    }
+
     this.dialog.open(ShowCodeDialog, {
       width: '50vw',
       enterAnimationDuration: 0,
       exitAnimationDuration: 0,
+      data: paramsToDisplay
     })
   }
 }
-