+++ src/app/model/params.ts
@@ -1,12 +1,28 @@
 // TODO: get available animation params for certain animation
+//  with eg. gerParam() function that is implemented by AnimatorInterface
 //  from animation file - PARAMS OBJ SHOULD BE GENERIC!
+export interface Params {
+  animation: string
+  autosize: boolean
+  background: string
+  color: string
+  fontsize: number
+  height: number
+  loop: string
+  text: string
+  textposition: string
+  width: number
+}
 
-export const DEFAULT_PARAMS = {
+export const DEFAULT_PARAMS: Params = {
   animation: "slide-right",
+  autosize: true,
   background: "url('assets/media/corollaSedan_1000x750.jpg')",
   color: "#666",
-  fontSize: "18",
+  fontsize: 18,
+  height: 750,
   loop: "true",
   text: "Wyprzedaż rocznika 2022",
-  textPosition: "bottom"
+  textposition: "bottom",
+  width: 1000,
 }