+++ changelog.md
@@ -1,3 +1,8 @@
+### Version 0.0.4
+- Dialog with instructions added
+### Version 0.0.3
+- Animated Banner params such custom color and custom background, width and height, can be edited
+- color picker component added
 ### Version 0.0.2
 - Animated Banner web component sample params preview
 ### Version 0.0.1